pr7

+++ pr-7/function.c
@@ -7,7 +7,7 @@ int sum(int a , int b){
 int subtraction (int a , int b){
     return a - b ; 
 }
-int division(int a , int b){
+float division(int a , int b){
     return a / b ; 
 }
 int multiplication(int a , int b){
